fix: accept neg lock ID for migration (#5727)
* accept 0 or neg lock ID

* add a comment explain neg number

(cherry picked from commit 294f3c0ab3f246f5d0f8e5ea8dce8536c47529f3)

+++ x/superfluid/keeper/msg_server_test.go
@@ -524,7 +524,7 @@ func (s *KeeperTestSuite) TestUnlockAndMigrateSharesToFullRangeConcentratedPosit
 	sender, err := sdk.AccAddressFromBech32(locks[0].Owner)
 	s.Require().NoError(err)
 	_, err = msgServer.UnlockAndMigrateSharesToFullRangeConcentratedPosition(sdk.WrapSDKContext(s.Ctx),
-		types.NewMsgUnlockAndMigrateSharesToFullRangeConcentratedPosition(sender, locks[0].ID, locks[0].Coins[0]))
+		types.NewMsgUnlockAndMigrateSharesToFullRangeConcentratedPosition(sender, int64(locks[0].ID), locks[0].Coins[0]))
 	s.Require().NoError(err)
 
 	// Asset event emitted
